python 3 changes

+++ src/watchdog/observers/winapi_common.py
@@ -18,6 +18,7 @@
 
 
 from watchdog.utils import platform
+from functools import reduce
 
 if platform.is_windows():
   import ctypes
@@ -130,7 +131,7 @@ if platform.is_windows():
     except WindowsError:
       return [], 0
     # get_FILE_NOTIFY_INFORMATION expects nBytes to be long.
-    return event_buffer.raw, long(nbytes.value)
+    return event_buffer.raw, int(nbytes.value)
 
 
   def create_io_completion_port():
